control_allocator: handle thrust allocation for tiltrotor VTOL properly
In the legancy mixing system, the mixer was getting fed with an abstract throttle setpoint
without any assosiated axis. Now we fed thrust in z for MC and thrust in x in FW, but keep
the same matrix for both flight phases. We can't always feed both thrust x and thrust z,
as that would be conflicting with the also fed tilt angle.
During transition thrust x is not set, only thrust z.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>